lofi-transfer/requirements.txt
torch>=2.0
torchaudio>=2.0
librosa>=0.10
soundfile>=0.12
numpy>=1.24
tqdm
pyyaml
tensorboard
